bright-fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-mcp-adapters
langchain-anthropic